Remove duplicated utils, config and constants files and add a link to each one in  each recipe readme
