--- cookbook-react-native/recipes/buttons/custom-hightlight-button/constants/colors.ts
@@ -1,7 +1,0 @@
-export const white = '#FFF';
-export const transparent = 'trasparent';
-export const  shadowColor = '#E5E5E5';
-export const primaryColor = '#1D8DFA';
-export const secondaryColor = '#0065C7';
-export const disabledBackground = '#F1F1F1';
-export const  disabledColor = '#939393';
