--- cookbook-react-native/recipes/buttons/custom-hightlight-button/constants/dimentions.ts
@@ -1,15 +1,0 @@
-import { scale } from "../utils/scaleUtils";
-
-export const BORDER_WIDTH = {
-  NONE: 0,
-  THIN: 0.5,
-  REGULAR: 1,
-  BOLD: 2
-};
-
-/* eslint-disable no-magic-numbers */
-export const SIZES = {
-  SMALL: scale(140),
-  MEDIUM: scale(160),
-  LARGE: scale(200)
-};
