--- cookbook-react-native/recipes/buttons/custom-hightlight-button/constants/interfaces.ts
@@ -1,7 +1,0 @@
-type Key = string | number;
-
-export type GenericObjectInterface<T> = {
-  [key in Key]: T;
-};
-
-export type NumberObject = GenericObjectInterface<number>;
