--- cookbook-react-native/recipes/buttons/custom-opacity-button/constants/colors.ts
@@ -1,10 +1,0 @@
-export const white: string = '#FFF';
-export const transparent: string = 'transparent';
-export const  shadowColor: string = '#E5E5E5';
-export const primaryColor: string = '#1D8DFA';
-export const secondaryColor: string = '#0065C7';
-export const disabledBackground: string = '#F1F1F1';
-export const  disabledColor: string = '#939393';
-export const green: string = '#53E69D';
-export const black: string = '#000';
-export const gray: string = '#939393';
