--- cookbook-react-native/recipes/buttons/custom-opacity-button/constants/interfaces.ts
@@ -1,7 +1,0 @@
-type Key = string | number;
-
-type GenericObjectInterface<T> = {
-  [key in Key]: T;
-};
-
-export type NumberObject = GenericObjectInterface<number>;
