--- cookbook-react-native/recipes/buttons/custom-opacity-button/constants/platform.ts
@@ -1,5 +1,0 @@
-import { Dimensions } from 'react-native';
-
-const windowDimensions = Dimensions.get('window');
-export const WINDOW_HEIGHT = windowDimensions.height;
-export const WINDOW_WIDTH = windowDimensions.width;
