+++ cookbook-react-native/recipes/buttons/custom-opacity-button/readme.md
@@ -1,0 +1,24 @@
+# Cookbook React Native Bootstrap: CustomOpacityButton Recipe
+
+If you want to add this recipe to a project that wasn't created with our [bootstrap](https://github.com/Wolox/wolmo-bootstrap-react-native), you will need to add the following [aliases](https://github.com/tleunen/babel-plugin-module-resolver#readme) and files under them to your project:
+
+## Utils files
+### alias `@utils`
+
+* [styleUtils.ts](https://github.com/Wolox/wolmo-bootstrap-react-native/blob/master/generators/app/templates/src/utils/styleUtils.ts)
+
+## Constants files
+### alias `@constants`
+
+* [colors.ts](https://github.com/Wolox/wolmo-bootstrap-react-native/blob/master/generators/app/templates/src/constants/colors.ts)
+* [platform.ts](https://github.com/Wolox/wolmo-bootstrap-react-native/blob/master/generators/app/templates/src/constants/platform.ts)
+
+## Interfaces files
+### alias `@interfaces`
+
+* [globalInterfaces.ts](https://github.com/Wolox/wolmo-bootstrap-react-native/blob/master/generators/app/templates/src/interfaces/globalInterfaces.ts)
+
+## Recipes
+### alias `@textsRecipes` but it would be recommended to put them in `@components`instead
+
+* [CustomText](https://github.com/Wolox/frontend-cookbook/master/cookbook-react-native/recipes/texts/custom-text#readme)
