+++ cookbook-react-native/recipes/buttons/custom-opacity-button/styles.ts
@@ -1,5 +1,5 @@
 import { StyleSheet } from 'react-native';
-import { transparent, black, green, gray, white } from './constants/colors';
+import { transparent, black, green, gray, white } from '@constants/colors';
 
 const iconSize = 20;
