--- cookbook-react-native/recipes/buttons/custom-opacity-button/utils/styleUtils.ts
@@ -1,5 +1,0 @@
-export function getCustomStyles(variants: string[], props: any, styles: any, stylePrefix: string = '') {
-  return variants
-    .map(variant => (props[variant] ? styles[`${variant}${stylePrefix}`] : null))
-    .filter(style => style !== null);
-}
