+++ cookbook-react-native/recipes/inputs/input-1/components/InputLabel/styles.ts
@@ -1,8 +1,7 @@
 import { StyleSheet } from 'react-native';
-
-import { transparent } from '../../constants/colors';
-import { SIZES } from '../../constants/fonts';
-import fonts from '../../config/fonts';
+import { transparent } from '@constants/colors';
+import { SIZES } from '@constants/fonts';
+import fonts from '@config/fonts';
 
 export default StyleSheet.create({
   label: {
