--- cookbook-react-native/recipes/inputs/input-1/constants/colors.ts
@@ -1,7 +1,0 @@
-export const primaryColor: string = '#288CCB';
-export const red: string = '#FF0000';
-export const transparent: string = 'transparent';
-export const white: string = '#FFF';
-export const black: string = '#000';
-export const gray: string = '#rgba(0, 0, 0, 0.38)';
-export const darkGray: string = '#363636';
