--- cookbook-react-native/recipes/inputs/input-1/constants/fonts.ts
@@ -1,22 +1,0 @@
-// FONTS
-// You can add here your custom fonts.
-// If you want to read more about naming and linking custom fonts,
-// this is the original post: https://medium.com/@mehran.khan/ultimate-guide-to-use-custom-fonts-in-react-native-77fcdf859cf4
-export const MONTSERRAT = 'Montserrat';
-// WEIGHTS
-export const REGULAR = 'Regular';
-export const SEMIBOLD = 'SemiBold';
-export const BOLD = 'Bold';
-// STYLES
-export const NORMAL = 'Normal';
-export const ITALIC = 'Italic';
-// SIZES
-export const SIZES = {
-  XXSMALL: 10,
-  XSMALL: 12,
-  SMALL: 14,
-  MEDIUM: 16,
-  XMEDIUM: 18,
-  BIG: 20,
-  XBIG: 36
-};
