--- cookbook-react-native/recipes/inputs/input-1/constants/interfaces.ts
@@ -1,7 +1,0 @@
-type Key = string | number;
-
-type GenericObjectInterface<T> = {
-  [key in Key]: T;
-};
-
-export type StringObject = GenericObjectInterface<string>;
