--- cookbook-react-native/recipes/screens/login/interfaces/navigation.ts
@@ -1,5 +1,0 @@
-import { RouteProp } from '@react-navigation/native';
-
-export interface DefaultNavOptions {
-  route: RouteProp<any, any>;
-}
