+++ cookbook-react-native/recipes/texts/custom-text/constants.ts
