--- cookbook-react-native/recipes/texts/custom-text/constants/colors.ts
@@ -1,7 +1,0 @@
-export const blue: string = '#288CCB';
-export const green: string = '#53E69D';
-export const red: string = '#FF0000';
-export const transparent: string = 'transparent';
-export const white: string = '#FFF';
-export const gray: string = '#rgba(0, 0, 0, 0.38)';
-export const darkGray: string = '#363636';
