--- cookbook-react-native/recipes/texts/custom-text/constants/interfaces.ts
@@ -1,9 +1,0 @@
-type Key = string | number;
-
-type GenericObjectInterface<T> = {
-  [key in Key]: T;
-};
-
-export type StringObject = GenericObjectInterface<string>;
-
-export type NumberObject = GenericObjectInterface<number>;
