--- cookbook-react-native/recipes/texts/custom-text/constants/platform.ts
@@ -1,9 +1,0 @@
-import { Platform, Dimensions } from 'react-native';
-
-export const ROOT = 'root';
-
-export const isAndroid = Platform.OS === 'android';
-
-const windowDimensions = Dimensions.get('window');
-export const WINDOW_HEIGHT = windowDimensions.height;
-export const WINDOW_WIDTH = windowDimensions.width;
