+++ cookbook-react-native/src/utils/fontUtils.ts
@@ -2,8 +2,7 @@ import { TextStyle } from 'react-native';
 import { isAndroid } from '@constants/platform';
 import { MONTSERRAT, REGULAR, SEMIBOLD, BOLD, NORMAL, ITALIC } from '@constants/fonts';
 import { StringObject } from '@interfaces/globalInterfaces';
-
-import { moderateScale } from './scalingUtils';
+import { moderateScale } from '@utils/scalingUtils';
 
 const REGULAR_WEIGHT: string = '400';
 const NORMAL_STYLE: string = 'normal';
