+++ cookbook-react-native/src/utils/validations/validateUtils.ts
